+++ apiserver/plane/api/serializers/__init__.py
@@ -44,6 +44,7 @@ from .issue import (
     IssueReactionSerializer,
     CommentReactionSerializer,
     IssueVoteSerializer,
+    IssuePublicSerializer,
 )
 
 from .module import (
